backend/requirements.txt
ace_tools==0.0
aiohttp==3.8.1
aiosignal==1.3.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.8
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
Deprecated==1.2.18
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-RESTful==0.3.10
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
gyp==0.1
h11==0.14.0
httpcore==1.0.8
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightgbm==4.6.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.9.0
mongoengine==0.28.2
more-itertools==8.10.0
multidict==6.4.3
netifaces==0.11.0
nose==1.3.7
numpy==1.25.1
nvidia-nccl-cu12==2.26.2
oauthlib==3.2.0
openai==1.73.0
packaging==24.0
pandas==2.0.3
pillow==11.1.0
playwright==1.50.0
pluggy==1.5.0
propcache==0.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycurl==7.44.1
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
PyGithub==2.6.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pymongo==4.6.3
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytils==0.4.1
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.6.0
tomli==2.2.1
tools==0.1.9
tqdm==4.67.1
Twisted==22.1.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.4.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
yarl==1.19.0
zipp==1.0.0
zope.interface==5.4.0
